+++ front/src/pages/Game/matchHistory.tsx
@@ -25,8 +25,8 @@ const MatchHistory = () => {
 
 	//<span onClick={add}> add</span>  <h1>Match history</h1>
 	return (
-		<div className="bg-white h-48 w-full bg-white/50 rounded-3xl p-4 sm:h-60 md:w-[59%] md:h-80 md:mt-px xl:w-[708px]">
-			<div className="h-10 text-3xl font-bold text-center">
+		<div className="bg-black/50 h-full w-full rounded-md shadow-md shadow-white">
+			<div className='h-[10%] flex justify-center items-center rounded-md shadow-lg bg-white/90'>
 				<h1>Match history</h1>
 			</div>
 			{/*<div style={{}} className="flex-col justify-center item-center overflow-y-auto h-28 w-full">
